Set default version to `0.0.0`
The previous default version was `__version__ = "0.1.0"`. Users of this
template may not want to begin at `0.1.0` because they may want to start
on a patch version first. Attempting to decrement a version with Hatch
will raise a `ValueError`. For example, attempting `hatch version 0.0.1`
raises "ValueError: Version `0.0.1` is not higher than the original
version `0.1.0`".

This commit will set the default version number to `0.0.0` so users of
this template can choose their initial version number.